Eliminiate uses of half and half4 in favor of float and float4. Word on the street is that most GPUs don't support half. More importantly, the Metal compiler 💩s with halfs and doesn't with floats. See issue 7941

+++ sources/Metal/Shaders/iTermCopyBackground.metal
@@ -27,9 +27,9 @@ iTermCopyBackgroundVertexShader(uint vertexID [[ vertex_id ]],
     return out;
 }
 
-fragment half4
+fragment float4
 iTermCopyBackgroundFragmentShader(iTermCopyBackgroundVertexFunctionOutput in [[stage_in]],
-                                  texture2d<half> texture [[ texture(iTermTextureIndexPrimary) ]]) {
+                                  texture2d<float> texture [[ texture(iTermTextureIndexPrimary) ]]) {
     constexpr sampler textureSampler (mag_filter::linear,
                                       min_filter::linear);
