+++ sources/Metal/Shaders/iTermTextShaderCommon.h
@@ -18,14 +18,14 @@ typedef struct {
     float2 backgroundTextureCoordinate;
     float4 textColor;
     float4 backgroundColor;
-    half4 underlineColor;
+    float4 underlineColor;
     int3 colorModelIndex;
     float2 textureOffset;  // Normalized offset in texture.
     float2 cellOffset;  // Coordinate of bottom left of cell in pixel coordinates. 0,0 is the bottom left of the screen.
     int underlineStyle;  // should draw an underline? For some stupid reason the compiler won't let me set the type as iTermMetalGlyphAttributesUnderline
     float2 viewportSize;  // size of viewport in pixels. TODO: see if I can avoid passing this to fragment function.
     float scale;  // 2 for retina, 1 for non-retina
-    half4 alphaVector;
+    float4 alphaVector;
 } iTermTextVertexFunctionOutput;
 
 typedef struct {
@@ -43,8 +43,8 @@ typedef struct {
 typedef struct {
     float4 clipSpacePosition [[position]];  // In vector function is normalized. In fragment function is in pixels, with a half pixel offset since it refers to the center of the pixel.
     float2 textureCoordinate;
-    half4 textColor;
-    half4 alphaVector;
+    float4 textColor;
+    float4 alphaVector;
 } iTermTextVertexFunctionOutputMonochrome;
 
 // Returns the weight in [0, 1] of underline for a pixel at `clipSpacePosition`.
@@ -64,7 +64,7 @@ float ComputeWeightOfUnderlineInverted(int underlineStyle,  // iTermMetalGlyphAt
                                        float2 textureCoordinate,
                                        float2 glyphSize,
                                        float2 cellSize,
-                                       texture2d<half> texture,
+                                       texture2d<float> texture,
                                        sampler textureSampler,
                                        float scale);
 
@@ -82,7 +82,7 @@ float ComputeWeightOfUnderlineRegular(int underlineStyle,  // iTermMetalGlyphAtt
                                       float2 textureCoordinate,
                                       float2 glyphSize,
                                       float2 cellSize,
-                                      texture2d<half> texture,
+                                      texture2d<float> texture,
                                        sampler textureSampler,
                                        float scale);
 
@@ -91,20 +91,20 @@ float ComputeWeightOfUnderlineRegular(int underlineStyle,  // iTermMetalGlyphAtt
 //
 // This simply implements bilinear interpolation using the sampler. See
 // iTermTextRenderer.mm for details on how the texture is structured.
-static inline half4 RemapColor(float4 scaledTextColor,  // scaledTextColor is the text color multiplied by float4(17).
+static inline float4 RemapColor(float4 scaledTextColor,  // scaledTextColor is the text color multiplied by float4(17).
                                float4 backgroundColor_in,
                                float4 bwColor_in,
-                               texture2d<half> models) {
+                               texture2d<float> models) {
     float4 bwColor = round(bwColor_in * 255) * 18 + 0.5;
     float4 backgroundColor = backgroundColor_in * 17 + 0.5;
 
     constexpr sampler s(coord::pixel,
                         filter::linear);
-    half r = models.sample(s, float2(bwColor.x + scaledTextColor.x,
-                                     backgroundColor.x)).x;
-    half g = models.sample(s, float2(bwColor.y + scaledTextColor.y,
-                                     backgroundColor.y)).x;
-    half b = models.sample(s, float2(bwColor.z + scaledTextColor.z,
-                                     backgroundColor.z)).x;
-    return half4(r, g, b, 1);
+    float r = models.sample(s, float2(bwColor.x + scaledTextColor.x,
+                                      backgroundColor.x)).x;
+    float g = models.sample(s, float2(bwColor.y + scaledTextColor.y,
+                                      backgroundColor.y)).x;
+    float b = models.sample(s, float2(bwColor.z + scaledTextColor.z,
+                                      backgroundColor.z)).x;
+    return float4(r, g, b, 1);
 }
